+++ sync-for-payables/src/codatsyncpayables/journals.py
@@ -30,7 +30,7 @@ class Journals:
         
         url = utils.generate_url(operations.CreateJournalRequest, base_url, '/companies/{companyId}/connections/{connectionId}/push/journals', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "journal_prototype", False, True, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, operations.CreateJournalRequest, "journal_prototype", False, True, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         query_params = utils.get_query_params(operations.CreateJournalRequest, request)