+++ sync-for-payables/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [python v3.0.0] sync-for-payables
 ### Releases
 - [PyPI v3.0.0] https://pypi.org/project/codat-sync-for-payables/3.0.0 - sync-for-payables
+
+## 2023-12-14 14:08:55
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.126.3 (2.214.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.1.0] sync-for-payables
+### Releases
+- [PyPI v3.1.0] https://pypi.org/project/codat-sync-for-payables/3.1.0 - sync-for-payables